refactor(infra): consolidate home path helpers
