+++ src/infra/home-dir.ts
@@ -130,16 +130,15 @@ export function resolveHomeRelativePath(
   return path.resolve(trimmed);
 }
 
-/**
- * Backward-compatible alias for resolving user paths against the effective home.
- *
- * @deprecated Use resolveHomeRelativePath.
- */
+/** Resolves a user path against the effective home, preserving blank inputs. */
 export function resolveUserPath(
   input: string,
   env: NodeJS.ProcessEnv = process.env,
   homedir: () => string = os.homedir,
 ): string {
+  if (!input) {
+    return "";
+  }
   return resolveHomeRelativePath(input, { env, homedir });
 }
